% Generated by build script
\name{datagov_dataset}
\alias{datagov_dataset}
\title{Get full details for a Data.gov dataset}
\usage{
datagov_dataset()
}
\arguments{
\item{id}{Dataset name or UUID (e.g. "annual-survey-of-manufactures")}
}
\value{tibble: one row with id, name, title, organization,}
\description{Get full details for a Data.gov dataset}
