% Generated by build script
\name{datagov_resources}
\alias{datagov_resources}
\title{Get resources (downloadable files) for a dataset}
\usage{
datagov_resources()
}
\arguments{
\item{id}{Dataset name or UUID}
}
\value{tibble: id, name, format, url, description, created (POSIXct), size}
\description{Get resources (downloadable files) for a dataset}
