% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagov.R
\name{datagov_tags}
\alias{datagov_tags}
\title{List Data.gov tags}
\usage{
datagov_tags(query = NULL, limit = 100)
}
\arguments{
\item{query}{Optional search term}

\item{limit}{Max results (default 100)}
}
\value{
tibble: id, name
}
\description{
Returns tags used to classify datasets.
}
