% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagov.R
\name{datagov_dataset}
\alias{datagov_dataset}
\title{Get full details for a Data.gov dataset}
\usage{
datagov_dataset(id)
}
\arguments{
\item{id}{Dataset name or UUID (e.g. "annual-survey-of-manufactures")}
}
\value{
tibble: one row with id, name, title, organization,
num_resources, metadata_created, notes
}
\description{
Returns metadata and resource list for a single dataset.
}
